moves all forms to the ui app
