+++ learningresources/tests/test_forms.py
@@ -3,7 +3,7 @@ Tests for learningresources forms.
 """
 from __future__ import unicode_literals
 
-from learningresources.forms import CourseForm
+from ui.forms import CourseForm
 from learningresources.models import Repository, Course
 
 from .base import LoreTestCase
